error 404 implementado

+++ src/main/java/br/com/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -58,25 +59,37 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalhesTopicoDto detalhar(@PathVariable Integer id) {
-		Topico topico = topicoRepository.getReferenceById(id);
-		return new DetalhesTopicoDto(topico);
+	public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Integer id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {			
+			return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Integer id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-		Topico topico = form.atualizar(id, topicoRepository);
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {			
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
 		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Integer id) {
-		topicoRepository.deleteById(id);
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {			
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
 		
-		return ResponseEntity.ok().build();
+		return ResponseEntity.notFound().build();
 	}
 	
 }
